Fix handlers on reconnect
In my eagerness to test the handlers, I failed to recognise I had
broken my reconnect logic! We need to pass the local mpd.Client into
handlers so that when the mpd connection drops and Client reconnects,
the new client is made available to the handlers.

Without this fix, after a reconnect, handlers would panic as *Client.C
would be null pointer.