:bug: Fix ocassional playlist fetch with bad start index
Ocassionally, we can't determine the current playlist size (the MPD
query fails), so we fetch a default slice of the playlist. This default
was start/end 0/0 which is an invalid range, it should be either -1/-1
or 0/n where n is greater than 0.

Use a sensible default of 0/max, so if we don't know the playlist
size, we fetch a manageable and valid range.